requirements.txt
Django==3.2
django-filter==2.4.0
django-rest-registration==0.6.0
django-seed==0.2.2
djangorestframework==3.12.4
